+++ requirements.txt
@@ -1,13 +1,17 @@
+alembic==1.13.1
 blinker==1.6.3
 click==8.1.7
 colorama==0.4.6
 Flask==3.0.0
+Flask-Migrate==4.0.5
 Flask-SQLAlchemy==3.1.1
 greenlet==3.0.3
 importlib-metadata==6.8.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
+Mako==1.3.2
 MarkupSafe==2.1.3
+python-dotenv==1.0.1
 SQLAlchemy==2.0.27
 typing_extensions==4.9.0
 Werkzeug==3.0.1
