--- site/app.py
@@ -1,24 +1,0 @@
-from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy
-
-
-app = Flask(__name__)
-
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///dev_db.sqlite3'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-db = SQLAlchemy(app)
-
-
-class Quote(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    text = db.Column(db.String(256), nullable=False)
-
-
-@app.route('/')
-def hello():
-    quotes = Quote.query.all()
-    return render_template('index.html', quotes=quotes)
-
-if __name__ == '__main__':
-    app.run(debug=True)
