+++ wisebook/database/quote_model.py
@@ -1,0 +1,6 @@
+from wisebook import db
+
+
+class Quote(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    text = db.Column(db.String(256), nullable=False)
