+++ wisebook/settings.py
@@ -1,0 +1,7 @@
+import os
+
+
+class Config(object):
+    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URI')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    SECRET_KEY = os.getenv('SECRET_KEY')
