+++ wisebook/templates/footer.html
